render 함수 구현
