+++ src/createElement.ts
@@ -1,20 +1,24 @@
+export interface VDom {
+  type: string;
+  props: Record<string, any>;
+  key?: string;
+}
+
 export function createElement(
   type: string,
   props: Record<string, any> | null,
   key?: string
-) {
-  const element = document.createElement(type);
-  if (key) element.setAttribute("key", key);
-
-  if (props) {
-    Object.entries(props).forEach(([k, v]) => {
-      if (k === "children") {
-        if (typeof v === "string") {
-          element.appendChild(document.createTextNode(v));
-        } else element.appendChild(createElement(v.type, v.props));
-      }
-    });
-  }
-
-  return element;
+): VDom {
+  const children = props?.children
+    ? Array.isArray(props.children)
+      ? props.children
+      : [props.children]
+    : [];
+  return {
+    type,
+    props: {
+      children: children.length > 1 ? children : children[0],
+    },
+    key,
+  };
 }
